Refactor folder structure for commands and testing
